pyproject.toml
[project]
name = "task-manager"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.1",
    "django-filter>=25.1",
    "django-hosts>=6.0",
    "django-tenants>=3.8.0",
    "djangorestframework>=3.16.0",
    "dotenv>=0.9.9",
    "markdown>=3.8",
    "psycopg2-binary>=2.9.10",
]

[dependency-groups]
dev = [
    "ruff>=0.11.12",
]
